+++ dst/app/routes/reservations.js
@@ -26,6 +26,8 @@ reservationsRouter.use(authentication_1.default);
  * 予約検索
  */
 reservationsRouter.get('', permitScopes_1.default(['admin', 'reservations', 'reservations.read-only']), ...[
+    express_validator_1.query('$projection.*')
+        .toInt(),
     express_validator_1.query('limit')
         .optional()
         .isInt()
@@ -76,11 +78,21 @@ reservationsRouter.get('', permitScopes_1.default(['admin', 'reservations', 'res
         .toBoolean()
 ], validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
+        const countDocuments = req.query.countDocuments === '1';
         const reservationRepo = new chevre.repository.Reservation(mongoose.connection);
         const searchConditions = Object.assign(Object.assign({}, req.query), { 
-            // tslint:disable-next-line:no-magic-numbers no-single-line-block-comment
-            limit: (req.query.limit !== undefined) ? Math.min(req.query.limit, 100) : 100, page: (req.query.page !== undefined) ? Math.max(req.query.page, 1) : 1, sort: { bookingTime: chevre.factory.sortType.Descending } });
-        const reservations = yield reservationRepo.search(searchConditions);
+            // tslint:disable-next-line:no-magic-numbers
+            limit: (req.query.limit !== undefined) ? Math.min(req.query.limit, 100) : 100, page: (req.query.page !== undefined) ? Math.max(req.query.page, 1) : 1, sort: (typeof req.query.sort === 'object' && req.query.sort !== undefined && req.query.sort !== null)
+                ? req.query.sort
+                : { bookingTime: chevre.factory.sortType.Descending } });
+        // projectionの指定があれば適用する
+        const projection = (req.query.$projection !== undefined && req.query.$projection !== null)
+            ? Object.assign({}, req.query.$projection) : undefined;
+        const reservations = yield reservationRepo.search(searchConditions, projection);
+        if (countDocuments) {
+            const totalCount = yield reservationRepo.count(searchConditions);
+            res.set('X-Total-Count', totalCount.toString());
+        }
         res.json(reservations);
     }
     catch (error) {
@@ -366,13 +378,78 @@ reservationsRouter.put('/eventReservation/screeningEvent/:id/checkedIn', permitS
         next(error);
     }
 }));
-reservationsRouter.put('/eventReservation/screeningEvent/:id/attended', permitScopes_1.default(['admin', 'reservations.attended']), validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+reservationsRouter.put('/eventReservation/screeningEvent/:id/attended', permitScopes_1.default(['admin', 'reservations.attended']), validator_1.default, 
+// tslint:disable-next-line:max-func-body-length
+(req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+    var _a, _b, _c, _d, _e;
     try {
+        const actionRepo = new chevre.repository.Action(mongoose.connection);
+        const projectRepo = new chevre.repository.Project(mongoose.connection);
         const reservationRepo = new chevre.repository.Reservation(mongoose.connection);
         const taskRepo = new chevre.repository.Task(mongoose.connection);
-        const reservation = yield reservationRepo.attend({
-            id: req.params.id
-        });
+        let reservation = yield reservationRepo.findById({ id: req.params.id });
+        const project = yield projectRepo.findById({ id: reservation.project.id });
+        // UseActionを作成する
+        const actionAttributes = Object.assign({ project: reservation.project, typeOf: chevre.factory.actionType.UseAction, agent: Object.assign({ typeOf: 'Person' }, req.body.agent), instrument: Object.assign({}, (typeof ((_a = req.body.instrument) === null || _a === void 0 ? void 0 : _a.token) === 'string')
+                ? { token: req.body.instrument.token }
+                : undefined), 
+            // どの予約を
+            object: [reservation] }, (typeof ((_b = req.body.location) === null || _b === void 0 ? void 0 : _b.identifier) === 'string')
+            ? {
+                location: {
+                    typeOf: chevre.factory.placeType.Place,
+                    identifier: req.body.location.identifier
+                }
+            }
+            : undefined
+        // purpose: params.purpose
+        );
+        let action = yield actionRepo.start(actionAttributes);
+        try {
+            reservation = (yield reservationRepo.attend({ id: reservation.id }));
+            // 使用日時がなければ追加
+            if (((_c = reservation.reservedTicket) === null || _c === void 0 ? void 0 : _c.dateUsed) === undefined) {
+                yield reservationRepo.reservationModel.findByIdAndUpdate(reservation.id, { 'reservedTicket.dateUsed': new Date() }, { new: true })
+                    .exec();
+            }
+        }
+        catch (error) {
+            // actionにエラー結果を追加
+            try {
+                const actionError = Object.assign(Object.assign({}, error), { message: error.message, name: error.name });
+                yield actionRepo.giveUp({ typeOf: action.typeOf, id: action.id, error: actionError });
+            }
+            catch (__) {
+                // 失敗したら仕方ない
+            }
+            throw error;
+        }
+        // アクション完了
+        action = yield actionRepo.complete({ typeOf: action.typeOf, id: action.id, result: {} });
+        const tasks = [];
+        // アクション通知タスク作成
+        const informAction = (_e = (_d = project.settings) === null || _d === void 0 ? void 0 : _d.onActionStatusChanged) === null || _e === void 0 ? void 0 : _e.informAction;
+        if (Array.isArray(informAction)) {
+            informAction.forEach((informParams) => {
+                const triggerWebhookTask = {
+                    project: action.project,
+                    name: chevre.factory.taskName.TriggerWebhook,
+                    status: chevre.factory.taskStatus.Ready,
+                    runsAt: new Date(),
+                    remainingNumberOfTries: 3,
+                    numberOfTried: 0,
+                    executionResults: [],
+                    data: {
+                        project: action.project,
+                        typeOf: chevre.factory.actionType.InformAction,
+                        agent: action.project,
+                        recipient: Object.assign({ typeOf: 'Person' }, informParams.recipient),
+                        object: action
+                    }
+                };
+                tasks.push(triggerWebhookTask);
+            });
+        }
         const aggregateTask = {
             project: reservation.project,
             name: chevre.factory.taskName.AggregateScreeningEvent,
@@ -386,9 +463,13 @@ reservationsRouter.put('/eventReservation/screeningEvent/:id/attended', permitSc
                 id: reservation.reservationFor.id
             }
         };
-        yield taskRepo.save(aggregateTask);
-        res.status(http_status_1.NO_CONTENT)
-            .end();
+        tasks.push(aggregateTask);
+        if (tasks.length > 0) {
+            yield taskRepo.saveMany(tasks);
+        }
+        // res.status(NO_CONTENT)
+        //     .end();
+        res.json({ id: action.id });
     }
     catch (error) {
         next(error);